# src/controllers/orders.controller.js
import {
  createOrder,
  getOrderById,
  getOrders,
  updateOrder,
  updateStatusOrder,
} from '../services/orders/orders.service.js'

export async function createOrderController(req, res, next) {
  const idUser = req.user.id_user
  const data = req.body

  try {
    const result = await createOrder(idUser, data)

    res.status(201).json({
      success: true,
      message: 'Pedido creado correctamente',
      order: result,
    })
  } catch (error) {
    next(error)
  }
}

export async function getOrdersController(req, res, next) {
  const idUser = req.user.id_user
  const data = req.query

  try {
    const result = await getOrders(idUser, data)

    res.status(200).json({
      success: true,
      orders: result,
    })
  } catch (error) {
    next(error)
  }
}

export async function getOrderByIdController(req, res, next) {
  const idUser = req.user.id_user
  const idOrder = req.params.id

  try {
    const result = await getOrderById(idUser, idOrder)

    res.status(200).json({
      success: true,
      order: result,
    })
  } catch (error) {
    next(error)
  }
}

export async function statusUpdateOrderController(req, res, next) {
  const idUser = req.user.id_user
  const idOrder = req.params.id
  const status = req.body

  try {
    const result = await updateStatusOrder(idUser, idOrder, status)

    res.status(200).json({
      success: true,
      message: 'Estado del pedido actualizado correctamente',
      order: result,
    })
  } catch (error) {
    next(error)
  }
}

export async function updateOrderController(req, res, next) {
  const idUser = req.user.id_user
  const idOrder = req.params.id
  const data = req.body

  try {
    const result = await updateOrder(idUser, idOrder, data)

    res.status(200).json({
      success: true,
      message: 'Pedido actualizado correctamente',
      order: result,
    })
  } catch (error) {
    next(error)
  }
}